+++ tutorial/Tutorial.jsx
@@ -6,7 +6,7 @@ This program is free software: you can redistribute it and/or modify it under th
 
 import React from 'react';
 import ReactDOM from 'react-dom';
-import Tutorial from './tutorial.rendered.mdx';
+import Tutorial from './tutorial.mdx';
 // import ApiDoc from './ApiDoc';
 import './style.scss';
 import '@strudel.cycles/react/dist/style.css';
